requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
google-auth==2.25.2
openai==1.6.1
python-multipart==0.0.6
alembic==1.13.1
python-dotenv==1.0.0
requests==2.31.0